Advance Organizer file
This html file needs to be placed within a module on the table of contents. This file is divided in three layers, the first layer is the course title bar which contains a home button that will take students to the org level home page, also through Brightspace Valence API the course title and the overall grade of a student its been display.  The second layer is the week navigator, the week navigator activates the week on which the current date belongs also each week is linked to an advance organizer file of that selected week. Lastly the third layer retrieves each topic from a particular module and assigns an icon and a url and placed this in buttons creating a navigation bar. When the buttons of this navigation bar are clicked the url will be loaded on the iframe bellow the navigation bar.